requirements.txt
# Soccer Film Analysis - Requirements
# Python 3.10+ recommended

# ============================================
# Core Detection & Tracking (Open Source)
# ============================================
# Roboflow Sports - Soccer-specific detection models
# Install via: pip install git+https://github.com/roboflow/sports.git
roboflow>=1.1.0
inference>=0.18.0
supervision>=0.25.0
ultralytics>=8.3.0

# ============================================
# Computer Vision & Video Processing
# ============================================
opencv-python>=4.8.0
opencv-python-headless>=4.8.0  # For server deployments
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.22.0

# ============================================
# Machine Learning & Clustering
# ============================================
scikit-learn>=1.3.0
torch>=2.0.0  # For GPU acceleration
torchvision>=0.15.0
umap-learn>=0.5.0  # For team classification

# ============================================
# OCR for Jersey Numbers
# ============================================
easyocr>=1.7.0
paddleocr>=2.7.0  # Alternative OCR option

# ============================================
# Database
# ============================================
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations

# ============================================
# GUI
# ============================================
PyQt6>=6.5.0
PyQt6-Qt6>=6.5.0
PyQt6-sip>=13.6.0

# ============================================
# Data Analysis & Visualization
# ============================================
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.18.0
mplsoccer>=1.2.0  # Soccer-specific visualizations

# ============================================
# Reporting
# ============================================
jinja2>=3.1.0
weasyprint>=60.0  # PDF generation
openpyxl>=3.1.0  # Excel export
python-docx>=1.0.0  # Word documents

# ============================================
# Configuration & Utilities
# ============================================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0
click>=8.1.0  # CLI
rich>=13.7.0  # Beautiful terminal output
tqdm>=4.66.0  # Progress bars

# ============================================
# Logging & Debugging
# ============================================
loguru>=0.7.0
colorlog>=6.8.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ============================================
# Development
# ============================================
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
